+++ packages/frontend-components/src/components/index.ts
@@ -36,3 +36,4 @@ export {
 export { DraggableGridLayout } from "./DraggableGridLayout/DraggableGridLayout";
 export { ExpandView } from "./Expand/ExpandView";
 export { Expand } from "./Expand/Expand";
+export { OverviewCard } from "./OverviewCard";